+++ npm-packages/shared-utils/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @pulse-editor/shared-utils
 
+## 0.1.1-alpha.6
+
+### Patch Changes
+
+- Update rollup settings
+
+## 0.1.1-alpha.5
+
+### Patch Changes
+
+- Update rollup settings
+
+## 0.1.1-alpha.4
+
+### Patch Changes
+
+- Configure rollup settings
+
+## 0.1.1-alpha.3
+
+### Patch Changes
+
+- Revert preserveModules settings
+
+## 0.1.1-alpha.2
+
+### Patch Changes
+
+- Add both CJS and ESM build
+
 ## 0.1.1-beta.1
 
 ### Patch Changes